Handle rotation in overlays
Before this patch, the rotation was automatically handled by Android.
It was working, but had some limitations, such as the impossibility to change
the layout depending on the orientation for instance.

Now, an overlay can be flag as "recreate on rotation", which will destroy the
overlay, update the context with the correct configuration, and show back the
overlay to the user with this corrected context.

This is a huge beneficial change, even if this current patch only contains
the orientation update, as it allows to easily add more configuration corrections,
such as dynamic theming colors (day/night transition, Material3 dynamic colors)